Extract completed download file to separate class.
